@@ -1,5 +1,5 @@
-from .application import Application, Applications
-from .computer import Computer, Computers
+from .application import Application, Versions
+from .computer import Computer
 from .config import ProgramConfig
 from .data import Data
 from .status import ProgramStatus, State, Status
