@@ -1,10 +1,12 @@
 """Data structures for application/computer status."""
 
 import functools
+from dataclasses import dataclass, field
 
 import log
-import yorm
 
+from .application import Application
+from .computer import Computer
 from .timestamp import Timestamp
 
 
@@ -45,15 +47,12 @@ def log_stopping(func):
     return wrapped
 
 
-@yorm.attr(computer=yorm.types.String)
-@yorm.attr(timestamp=Timestamp)
-class State(yorm.types.AttributeDictionary):
+@dataclass
+class State:
     """Dictionary of computer state."""
 
-    def __init__(self, computer=None, timestamp=None):
-        super().__init__()
-        self.computer = computer
-        self.timestamp = timestamp or Timestamp()
+    computer: str
+    timestamp: Timestamp = field(default_factory=Timestamp)
 
     def __str__(self):
         return str(self.computer)
@@ -62,24 +61,13 @@ class State(yorm.types.AttributeDictionary):
         return str(self.computer).lower() < str(other.computer).lower()
 
 
-@yorm.attr(all=State)
-class StateList(yorm.types.SortedList):
-    """List of computer states for an application."""
-
-
-@yorm.attr(application=yorm.types.String)
-@yorm.attr(computers=StateList)
-@yorm.attr(next=yorm.types.NullableString)
-class Status(yorm.types.AttributeDictionary):
+@dataclass
+class Status:
     """Dictionary of computers using an application."""
 
-    def __init__(
-        self, application=None, computers=None, next=None
-    ):  # pylint: disable=redefined-builtin
-        super().__init__()
-        self.application = application
-        self.computers = computers or StateList()
-        self.next = next
+    application: str
+    computers: list[State] = field(default_factory=list)
+    next: str | None = None
 
     def __str__(self):
         return str(self.application)
@@ -88,20 +76,12 @@ class Status(yorm.types.AttributeDictionary):
         return str(self.application).lower() < str(other.application).lower()
 
 
-@yorm.attr(all=Status)
-class StatusList(yorm.types.SortedList):
-    """List of application statuses."""
-
-
-@yorm.attr(applications=StatusList)
-@yorm.attr(counter=yorm.types.Integer)
-class ProgramStatus(yorm.types.AttributeDictionary):
+@dataclass
+class ProgramStatus:
     """Dictionary of current program status."""
 
-    def __init__(self, applications=None, counter=0):
-        super().__init__()
-        self.applications = applications or StatusList()
-        self.counter = counter
+    applications: list[Status] = field(default_factory=list)
+    counter: int = 0
 
     def find(self, application):
         """Return the application status for an application."""
@@ -128,7 +108,7 @@ class ProgramStatus(yorm.types.AttributeDictionary):
                     # TODO: consider returning the computer instance?
                     return states[0].computer
 
-        log.debug("marked as started on: nothing")
+        log.debug(f"{application} marked as started on: nothing")
         return None
 
     @log_running
@@ -143,13 +123,13 @@ class ProgramStatus(yorm.types.AttributeDictionary):
         # Status not found, assume the application is not running
         return False
 
-    def queue(self, application, computer):
+    def queue(self, application: Application, computer: Computer):
         """Record an application as queued for launch on a computer."""
         status = self.find(application)
         status.next = computer.name
 
     @log_starting
-    def start(self, application, computer):
+    def start(self, application: Application, computer: Computer):
         """Record an application as running on a computer."""
         for status in self.applications:
             if status.application == application.name:
@@ -174,7 +154,7 @@ class ProgramStatus(yorm.types.AttributeDictionary):
             status.computers.append(state)
 
     @log_stopping
-    def stop(self, application, computer):
+    def stop(self, application: Application, computer: Computer):
         """Record an application as no longer running on a computer."""
         for status in self.applications:
             if status.application == application.name:
